+++ packages/site-demo/content/product/components/tabs/react/icon-tabs.tsx
@@ -1,31 +1,21 @@
 import { mdiBowl, mdiBreadSliceOutline, mdiSilverwareForkKnife } from '@lumx/icons';
-import { Tab, Tabs, Theme } from '@lumx/react';
+import { Tab, TabList, TabPanel, TabProvider } from '@lumx/react';
 import React, { useState } from 'react';
 
 export const App = ({ theme }: any) => {
     const [activeTab, setActiveTab] = useState(0);
-    const handleTabClick = ({ index }: any) => {
-        setActiveTab(index);
-    };
 
     return (
-        <Tabs
-            className={theme === Theme.dark && 'lumx-color-font-light-N'}
-            theme={theme}
-            activeTab={activeTab}
-            onTabClick={handleTabClick}
-        >
-            <Tab label="Tab 1" icon={mdiBowl} className>
-                <p className="lumx-spacing-padding-vertical-huge">Tab 1 content</p>
-            </Tab>
+        <TabProvider activeTabIndex={activeTab} onChange={setActiveTab}>
+            <TabList theme={theme} aria-label="Tab list">
+                <Tab label="Tab 1" icon={mdiBowl} />
+                <Tab label="Tab 2" icon={mdiBreadSliceOutline} />
+                <Tab label="Tab 3" icon={mdiSilverwareForkKnife} />
+            </TabList>
 
-            <Tab label="Tab 2" icon={mdiBreadSliceOutline}>
-                <p className="lumx-spacing-padding-vertical-huge">Tab 2 content</p>
-            </Tab>
-
-            <Tab label="Tab 3" icon={mdiSilverwareForkKnife}>
-                <p className="lumx-spacing-padding-vertical-huge">Tab 3 content</p>
-            </Tab>
-        </Tabs>
+            <TabPanel className="lumx-spacing-padding-vertical-huge">Tab 1 content</TabPanel>
+            <TabPanel className="lumx-spacing-padding-vertical-huge">Tab 2 content</TabPanel>
+            <TabPanel className="lumx-spacing-padding-vertical-huge">Tab 3 content</TabPanel>
+        </TabProvider>
     );
 };